Testing and Debugging
1. Fix issue with messages not appearing on left or right if it is too long
2. Fix issue where there is no confirmation to delete a reflection
3. Fix issue where counsellor dashboard doesn't update trailing icons

+++ lib/ui/student/journaling/journal_reflect/journal_reflect_view.dart
@@ -112,8 +112,39 @@ class _MyJournalReflectViewState extends State<JournalReflectView> {
       child: ListTile(
         trailing: IconButton(
           icon: const Icon(Icons.delete),
-          onPressed: () => controller.firestoreService
-              .deleteReflection(controller.date, reflection),
+          onPressed: () async {
+            bool? confirmDelete = await showDialog<bool>(
+              context: context,
+              builder: (BuildContext context) {
+                return AlertDialog(
+                  title: const Text('Confirm Deletion'),
+                  content: const Text(
+                      'Are you sure you want to delete this reflection?'),
+                  actions: <Widget>[
+                    TextButton(
+                      onPressed: () {
+                        Navigator.of(context)
+                            .pop(false); // User cancelled the action
+                      },
+                      child: const Text('Cancel'),
+                    ),
+                    TextButton(
+                      onPressed: () {
+                        Navigator.of(context)
+                            .pop(true); // User confirmed the deletion
+                      },
+                      child: const Text('Delete'),
+                    ),
+                  ],
+                );
+              },
+            );
+
+            if (confirmDelete == true) {
+              controller.firestoreService
+                  .deleteReflection(controller.date, reflection);
+            }
+          },
         ),
         title: Text(reflection,
             style: AppTextStyles.mediumBlackText,
